Thrid Pass
FM - Not sure how use the index of the canidate work what has been transformed into a dictionary. This used the dictionary part but just loops through the letters of the list words that have been ordered in desendingly. If all the letters are found then it reutens the word.

Signed-off-by: Frank McAuley <frank.mcauley@gmail.com>